Set FD_CLOEXEC for ipmi driver device file.
If a privileged process, which uses this library, spawns a
non-privileged child, the child can inherit open file descriptor to ipmi
driver. This may become a significant both safety and security
vulnerability, if access to ipmi device file is not intended (For
details see [1]).

Another possibility for Linux is to use O_CLOEXEC flag, which is
considered to be thread safe. But F_CLOEXEC is more portable.

1. https://cwe.mitre.org/data/definitions/403.html