Warning about NUL
The issues described herein was the cause of a multi-hour debugging
session after *thousands* of consoles were locked into the wrong baud
rate in a datacenter.

Happy to change the wording, but I feel like some kind of "this is not a
theoretical problem" warning needs to be in there.  :)