ipmi-sel: make --date-range more granular
According to Timestamp section of the IPMI specification[0], the SEL
entries have a timestamp resolution of one second. Therefore it is
possible to make the command line argument "--date-range" of ipmi-sel
more granular.

This patch is adding the usage of (almost)ISO 8601[1] style datetime strings
in the "--date-range" argument. The timezone info part of the standard
is not implemented since IPMI does not have proper support for it and
always assumes the users localtime.

So the option now also accepts datetime format strings in the form

  %Y-%m-%dT%H:%M:%S

e.g. in order to represent the datetime October 14th 2022, 09:05 AM,
one would write:

  2022-10-14T09:05:00

A new example usage of the "--date-range" leveraging this would be

  $ ipmi-sel --date-range=2022-10-14T09:05:00-2022-10-15T19:05:00

in order to show SEL records between Oct 14th 09:05 and Oct 15th 19:05
of the year 2022.

[0]
https://www.intel.com/content/www/us/en/products/docs/servers/ipmi/ipmi-second-gen-interface-spec-v2-rev1-1.html
[1] https://www.iso.org/obp/ui/#iso:std:iso:8601:-1:ed-1:v1:en

Closes: #57
Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>