configure.ac: Use AC_USE_SYSTEM_EXTENSIONS instead of appending -D_GNU_SOURCE
This achieves the same effect as it'll use -D_GNU_SOURCE when possible
but it ensures we never forget it in an individual Makefile.am.

AC_USE_SYSTEM_EXTENSIONS was introduced in autoconf 2.60, so we bump
the requirement from 2.57->2.60. Given the comment mentions "modern
autoconf and automake", we could really justify bumping this to 2.69
which itself is not exactly brand new, if we want.

Signed-off-by: Sam James <sam@gentoo.org>