Mention OpenIPMI in opensesspriv workaround doc
The "opensesspriv" workaround is needed also for the OpenIPMI LAN BMC
code (the ipmi_sim IPMI LAN BMC simulator and probably also the ipmilan
IPMI LAN to system interface converter, although I have not tested the
latter). The error message when the workaround is not used is different
from the other documented cases though:
"privilege level cannot be obtained for this user". Document this in
the man pages.